move some components into a more sensible place
