+++ frontend/src/features/card/card.tsx
@@ -21,9 +21,9 @@ import styled from "styled-components";
 import { RootState } from "@/app/store";
 import { SearchQuery, useAppDispatch, useAppSelector } from "@/common/types";
 import { CardDocument } from "@/common/types";
+import { Spinner } from "@/components/spinner";
 import { setSelectedCardAndShowModal } from "@/features/modals/modalsSlice";
 import { selectCardDocumentByIdentifier } from "@/features/search/cardDocumentsSlice";
-import { Spinner } from "@/features/ui/spinner";
 
 const HiddenImage = styled(Image)`
   z-index: 0;
