+++ frontend/src/features/export/export.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import Dropdown from "react-bootstrap/Dropdown";
 
+import { RightPaddedIcon } from "@/components/icon";
 import { ExportDecklist } from "@/features/export/exportDecklist";
 import { ExportImages } from "@/features/export/exportImages";
 import { ExportXML } from "@/features/export/exportXML";
-import { RightPaddedIcon } from "@/features/ui/styledComponents";
 
 export function Export() {
   return (
