+++ frontend/src/features/import/importXML.tsx
@@ -28,6 +28,7 @@ import { TextFileDropzone } from "@/common/dropzone";
 import { processPrefix } from "@/common/processing";
 import { useAppDispatch, useAppSelector } from "@/common/types";
 import { Cardstock, SlotProjectMembers } from "@/common/types";
+import { RightPaddedIcon } from "@/components/icon";
 import {
   setCardstock,
   setFoil,
@@ -38,7 +39,6 @@ import {
   selectProjectSize,
   setSelectedCardback,
 } from "@/features/project/projectSlice";
-import { RightPaddedIcon } from "@/features/ui/styledComponents";
 
 export function ImportXML() {
   const dispatch = useAppDispatch();
