+++ frontend/src/features/modals/cardDetailedViewModal.tsx
@@ -5,36 +5,24 @@
  */
 
 import { saveAs } from "file-saver";
-import React, { memo, useRef, useState } from "react";
+import React, { memo } from "react";
 import Badge from "react-bootstrap/Badge";
 import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
-import OverlayTrigger from "react-bootstrap/OverlayTrigger";
 import Row from "react-bootstrap/Row";
 import Table from "react-bootstrap/Table";
-import Tooltip from "react-bootstrap/Tooltip";
-import styled from "styled-components";
 
 import { api, useGetLanguagesQuery } from "@/app/api";
 import { base64StringToBlob } from "@/common/processing";
 import { CardDocument } from "@/common/types";
 import { imageSizeToMBString } from "@/common/utils";
+import { ClickToCopy } from "@/components/clickToCopy";
+import DisableSSR from "@/components/disableSSR";
+import { Spinner } from "@/components/spinner";
 import {
   MemoizedCardImage,
   MemoizedCardProportionWrapper,
 } from "@/features/card/card";
-import DisableSSR from "@/features/ui/disableSSR";
-import { Spinner } from "@/features/ui/spinner";
-
-const ClickToCopyIdentifier = styled.code`
-  user-select: none;
-  outline: solid 1px #ffffff00;
-  transition: outline 0.2s ease-in-out;
-  &:hover {
-    outline-color: #ffffffff;
-    cursor: pointer;
-  }
-`;
 
 interface CardDetailedViewProps {
   cardDocument: CardDocument;
@@ -57,13 +45,6 @@ export function CardDetailedViewModal({
   const languageNameByCode = Object.fromEntries(
     (getLanguagesQuery.data ?? []).map((row) => [row.code, row.name])
   );
-
-  const [copied, setCopied] = useState<boolean>(false);
-  const copyIdentifier = () => {
-    navigator.clipboard.writeText(cardDocument.identifier);
-    setCopied(true);
-    setTimeout(() => setCopied(false), 1000);
-  };
 
   const downloadImage = async () => {
     const response = await triggerFn(cardDocument.identifier);
@@ -144,19 +125,7 @@ export function CardDetailedViewModal({
                       <b>Identifier</b>
                     </td>
                     <td>
-                      <OverlayTrigger
-                        defaultShow={false}
-                        placement="top"
-                        overlay={
-                          <Tooltip id="image-identifier">
-                            {copied ? "Copied!" : "Click to copy"}
-                          </Tooltip>
-                        }
-                      >
-                        <ClickToCopyIdentifier onClick={copyIdentifier}>
-                          {cardDocument.identifier}
-                        </ClickToCopyIdentifier>
-                      </OverlayTrigger>
+                      <ClickToCopy text={cardDocument.identifier} />
                     </td>
                   </tr>
                   <tr>
